Reorganized some files and cleaned up comments
