+++ app/src/main/java/dev/tavarus/boardgamelogger/data/GamesDatabase.kt
@@ -3,10 +3,10 @@ package dev.tavarus.boardgamelogger.data
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import dev.tavarus.boardgamelogger.data.apimodels.DBPlay
-import dev.tavarus.boardgamelogger.data.apimodels.DBScore
+import dev.tavarus.boardgamelogger.data.apimodels.play.DBPlay
+import dev.tavarus.boardgamelogger.data.apimodels.play.DBScore
 import dev.tavarus.boardgamelogger.data.apimodels.Player
-import dev.tavarus.boardgamelogger.data.apimodels.ScoreConverter
+import dev.tavarus.boardgamelogger.data.apimodels.play.ScoreConverter
 
 @Database(entities = [DBPlay::class, DBScore::class, Player::class], version = 1)
 @TypeConverters(ScoreConverter::class)
